pyproject.toml
[tool.poetry]
name = "mlops"
version = "0.1.0"
description = "This is a project for the MLOps course"
authors = ["Alexander Yaskevich"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "mlops"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
numpy = "^1.26.0"
pandas = "^2.1.1"
scikit-learn = "^1.3.1"
joblib = "^1.3.2"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
mlflow = "^2.8.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
dvc = "^3.30.3"
dvc-gdrive = "^2.20.0"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
src_paths = ["mlopscourse"]
profile = "black"
line_length = 88
lines_after_imports = 2

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
